chore: rename directory
